ai_pdf_qa_tool/requirements.txt
fastapi 
uvicorn 
python-multipart 
jinja2 
pypdf 
openai 
python-dotenv 
requests